Unify sizes of `amdsmi_power_info_t`
Sizes are used inconsistently.  This causes tools to not show
N/A when they should.  Make them unified.

+++ include/amd_smi/amdsmi.h
@@ -1090,12 +1090,12 @@ typedef struct {
  * @cond @tag{gpu_bm_linux} @tag{host} @endcond
  */
 typedef struct {
-    uint64_t socket_power;          //!< Socket power in W {@linux_bm}, uW {@host}
+    uint32_t socket_power;          //!< Socket power in W {@linux_bm}, uW {@host}
     uint32_t current_socket_power;  //!< Current socket power in W {@linux_bm}, Linux only, Mi 300+ Series cards
     uint32_t average_socket_power;  //!< Average socket power in W {@linux_bm}, Linux only, Navi + Mi 200 and earlier Series cards
-    uint64_t gfx_voltage;           //!< GFX voltage measurement in mV {@linux_bm} or V {@host}
-    uint64_t soc_voltage;           //!< SOC voltage measurement in mV {@linux_bm} or V {@host}
-    uint64_t mem_voltage;           //!< MEM voltage measurement in mV {@linux_bm} or V {@host}
+    uint32_t gfx_voltage;           //!< GFX voltage measurement in mV {@linux_bm} or V {@host}
+    uint32_t soc_voltage;           //!< SOC voltage measurement in mV {@linux_bm} or V {@host}
+    uint32_t mem_voltage;           //!< MEM voltage measurement in mV {@linux_bm} or V {@host}
     uint32_t power_limit;           //!< The power limit in W {@linux_bm}, Linux only
     uint64_t reserved[18];
 } amdsmi_power_info_t;
